==> 01_make_stock_fundamental.ipynb <==
import pandas as pd
from pykrx import stock

ticker = '005930'

latest_date = '20230203'
start_date = latest_date
end_date = latest_date

df_stock = stock.get_market_fundamental(latest_date, latest_date, ticker)
df_stock

df_stock.index.name =None
df_stock

df_stock.index = [ticker]
df_stock

name = stock.get_market_ticker_name(ticker)
name

df_stock.insert(0, 'Name', name)
df_stock

def get_stock_fundamental(date, ticker):
    # 펀더멘탈 정보 읽어 오기
    df_stock = stock.get_market_fundamental(date, date, ticker)
    # 인덱스 이름 제거
    df_stock.index.name = None
    # 인덱스를 티커로 변경
    df_stock.index = [ticker]
    # 이름 읽어 오기
    name = stock.get_market_ticker_name(ticker)
    # 첫 번째 열에 삽입하기
    df_stock.insert(0, 'Name', name)
    # 주식 데이터 리턴
    return df_stock

df_stock = get_stock_fundamental('20230203', '005930')
df_stock
