==> 10_PER_PBR_기준_코스피_저평가우량주_선정하기.ipynb <==
import pandas as pd
from pykrx import stock

target_date = '20230213'

"""
코스피 상장 주식 펀더멘탈 정보 읽어 오기
'"""
df_f = stock.get_market_fundamental_by_ticker(target_date, market='KOSPI')

df_f

"""
PER 0 이하 제거
"""
df_f.drop(df_f[df_f.PER <= 0].index, inplace=True)

df_f.drop(df_f[df_f.PBR <= 0].index, inplace=True)

df_f

"""
통계 정보 보기
"""
df_f.describe()

df = df_f

df['PER*PBR'] = df.PER * df.PBR

df

df['PER_PBR_RANK'] = df['PER*PBR'].rank()

df

df.sort_values(by='PER_PBR_RANK', inplace=True)

df

"""
이름 추가하기
"""
names = []
for ticker in df.index:
    try:
        name = stock.get_market_ticker_name(ticker)
        names.append(name)
    except Exception as e:
        print(f' exception happens while getting name of ticker = {ticker}')
        print(e)
        names.append('N/A')

df.insert(0, 'NAME', names)

df

df.head(20)

df_div = df.drop(df[df.DPS <= 0].index)

df_div.head(20)
