==> 20_TIGER_배당성장_ETF_CSV_파일_전처리.ipynb <==
TIGER 배당성장 ETF CSV 파일을 xlsx 파일로 저장

import pandas as pd

def convertTigerEtfCsvToXlsx(csvFilePath):
    try:
        df = pd.read_csv(csvFilePath, encoding='EUC-KR')

        """ No 컬럼 삭제 """
        df = df.drop('No', axis=1)

        """ 종목코드가 0인 행 삭제 """
        df = df[df.종목코드 != 0]

        """ 정수형 종목코드를 6자리 스트링으로 변환 """
        tickers = df.종목코드.to_list()
        tickers_new = []
        for ticker in tickers:
            """ 앞에 빈 자리는 0으로 채워서 6자리로 만든다. """
            tickers_new.append(str(ticker).zfill(6))
        df.종목코드 = tickers_new

        df.to_excel(csvFilePath.replace('.csv', '.xlsx'), index=False)
    except Exception as e:
        raise e

"""
파일 변환
"""
#csvFilePath = './data/PDF_DATA.csv'
csvFilePath = './data/PDF_DATA_20230524.csv'

try:
    convertTigerEtfCsvToXlsx(csvFilePath)
except Exception as e:
    print(e)

"""FIN"""

"""
패키지/모듈 만들기
"""

import kodivstock.tigercsvtoxlsx as tigercsvtoxlsx

xlsxFilePath = tigercsvtoxlsx.convertTigerEtfCsvToXlsx('./data/PDF_DATA.csv')
xlsxFilePath

import pandas as pd

df = pd.read_excel(xlsxFilePath)
df

"""
csv 파일 안의 문자 인코딩 방식 확인하기
"""
import chardet

with open('./data/PDF_DATA.csv', 'rb') as rawdata:
    result = chardet.detect(rawdata.read(10000))

# check what the character encoding might be
print(result)

"""
csv 파일 안의 문자 인코딩 방식 확인하기
"""
import chardet

def getEncoding(filePath):
    with open('./data/PDF_DATA.csv', 'rb') as rawdata:
        result = chardet.detect(rawdata.read(10000))

    # check what the character encoding might be
    return result['encoding']

encoding = getEncoding('./data/PDF_DATA.csv')
encoding

chardet.__version__

df = pd.read_csv('PDF_DATA.csv', encoding='EUC-KR')
df.head(3)

df.tail(3)

df = df[df.종목코드 != 0]
df.tail(3)

df.info()

tickers = df.종목코드.to_list()
tickers

"""
정수형 종목코드를 6자리 스트링으로 변환
앞에 빈 자리는 0으로 채워서 6자리로 만든다.
"""
tickers_new = []
for ticker in tickers:
    if ticker > 0:
        tickers_new.append(str(ticker).zfill(6))
tickers_new

df.종목코드 = tickers_new
df

df = df.drop('No', axis=1)
df.head(3)

df.info()

#df.to_csv('PDF_DATA.csv', index=False, encoding="utf-8-sig")
df.to_excel('PDF_DATA.xlsx', index=False, encoding="utf-8-sig")

from datetime import datetime, timedelta
def getLatestBusinessDay():
    yesterday = datetime.now() - timedelta(1)
    yesterday = datetime.strftime(yesterday, '%Y%m%d')
    return stock.get_nearest_business_day_in_a_week(yesterday)

latest_business_day = getLatestBusinessDay()
latest_business_day

"""
타이거 배당성장 ETF 50개 Fundamental 정보 수집
"""

tickers = ['005490', '011780'] # TEST ONLY
df_tiger_div_growth_50 = None
#start_time = time.time()
for idx, ticker in enumerate(tickers):
    print(f'{idx+1:03d}: {ticker}')
    df_stock_fundamental = stock.get_market_fundamental(latest_business_day, latest_business_day, ticker)
    display(df_stock_fundamental)
    df_stock_fundamental.index.name = None
    df_stock_fundamental.index = [ticker]
    if df_tiger_div_growth_50 is None:
        df_tiger_div_growth_50 = df_stock_fundamental
    else:
        df_tiger_div_growth_50 = pd.concat([df_tiger_div_growth_50, df_stock_fundamental])
    time.sleep(1)
#end_time = time.time()

df_tiger_div_growth_50

df_tiger_div_growth_50.index.name = None

df_tiger_div_growth_50

df = pd.read_excel('PDF_DATA.xls')
df
