==> 22_TIGER_배당성장_50종목_배당성장주_TOP_PICK.ipynb <==
import numpy as np
import pandas as pd
from scipy.signal import savgol_filter
from pykrx import stock
import matplotlib.pyplot as plt
import time
plt.rcParams['font.family'] = 'Malgun Gothic'

import kodivstock.divexcelanalyzer as dea

df = pd.read_excel('./data/PDF_DATA.xlsx')
df.head(3)

df_result = df.iloc[:50, :2]
df_result.head(3)

df_result.tail(3)

""" 배당 성장 분석하기 """

divExcelAnalyzer = dea.DivExcelAnalyzer('./data/tiger_divgrowth50_dps_data_2012_2021.xlsx')
tickers = divExcelAnalyzer.getTickers()
len(tickers)

""" 배당 성장 연수 구하기 """
div_growth_years = divExcelAnalyzer.getDivGrowthYears(tickers)
div_growth_years[:3]

df_result['Div Growth Years'] = div_growth_years
df_result.head(3)

df_result.tail(3)

div_growth_rates_7 = divExcelAnalyzer.getDivGrowthRates(tickers, 7)
div_growth_rates_7[:3]

df_result['Div Growth Rates(7)'] = div_growth_rates_7
df_result.head(3)

div_growth_rates_3 = divExcelAnalyzer.getDivGrowthRates(tickers, 3)
div_growth_rates_3[:3]

df_result['Div Growth Rates(3)'] = div_growth_rates_3
df_result.head(3)

""" 정수형 종목코드를 6자리 스트링으로 변환 """
tickers = df_result.종목코드.to_list()
print(len(tickers))
tickers_new = []
for ticker in tickers:
    """ 앞에 빈 자리는 0으로 채워서 6자리로 만든다. """
    tickers_new.append(str(ticker).zfill(6))
df_result.종목코드 = tickers_new
df_result

df_result.to_excel('./data/tiger_divgrowth50_div_analyze_result.xlsx', index=False )

"""
result 파일 분석
"""

df_result = pd.read_excel('./data/tiger_divgrowth50_div_analyze_result.xlsx')
df_result.head(3)

""" 정수형 종목코드를 6자리 스트링으로 변환 """
tickers = df_result.종목코드.to_list()
print(len(tickers))
tickers_new = []
for ticker in tickers:
    """ 앞에 빈 자리는 0으로 채워서 6자리로 만든다. """
    tickers_new.append(str(ticker).zfill(6))
df_result.종목코드 = tickers_new
df_result.head(3)

df_result.describe()

df_divgrowth_over_7 = df_result.loc[df_result['Div Growth Years'] >= 7]
df_divgrowth_over_7.head(10)

df_divgrowth_over_7.to_excel('./data/tiger_divgrowth50_7년이상배당성장.xlsx', index=False)

"""
7년 이상 배당 성장주를 7년 배당 성장률로 정렬 후 상위 20개 선택
"""
df_divgrowth_over_7_top20 = df_divgrowth_over_7.sort_values(['Div Growth Rates(7)'], ascending=False).head(20)
df_divgrowth_over_7_top20

"""
7년 이상 배당 성장주를 3년 배당 성장률로 정렬 후 상위 20개 선택
"""
df_divgrowth_over_7.sort_values(['Div Growth Rates(3)'], ascending=False).head(20)

"""
7년 이상 배당 성장주 중 최근 3년 배당 성장률이 7년 성장률보다 높은 종목
"""
df_divgrowth_accelerated = df_divgrowth_over_7.loc[df_divgrowthyears['Div Growth Rates(3)']
                                                 > df_divgrowthyears['Div Growth Rates(7)']]
df_divgrowth_accelerated

df_divgrowth_accelerated.to_excel('./data/tiger_divgrowth50_배당성장PICK.xlsx', index=False)

"""
FIN
"""
