==> 23_TIGER_배당성장_50종목_적정주가_계산하기.ipynb <==
import os
import numpy as np
import pandas as pd
from scipy.signal import savgol_filter
from pykrx import stock
import matplotlib.pyplot as plt
import time
plt.rcParams['font.family'] = 'Malgun Gothic'
from KorDivStockPricer import KorDivStockPricer

def makeDirIfNotExist(dir):
    try:
        if not os.path.exists(dir):
            os.makedirs(dir)
    except Exception as e:
        print(str(e))

OUT_DIR = os.path.join(os.getcwd(), 'output')
OUT_DIR

makeDirIfNotExist(OUT_DIR)

START_YEAR = 2012
END_YEAR = 2021

df = pd.read_excel('./data/tiger_divgrowth50_div_analyze_result.xlsx')
df.head(3)

df.tail(3)

#df_target = df.head(2)
df_target = df
pricingResults = []
for code in df_target.종목코드:
    code = str(code).zfill(6)
    print(f'code = {code}')
    stockPricer = KorDivStockPricer(code)
    stockPricer.getStockName()
    stockPricer.collectStockData()
    pricingResult = stockPricer.doStockPricing()
    pricingResults.append(pricingResult)
    stockPricer.savePriceDivChart()
    time.sleep(1)

len(pricingResults)

df_total = None
for pricingResult in pricingResults:
    #print(pricingResult)
    df_pricing_result = pd.DataFrame(pricingResult)
    if df_total is None:
        df_total = df_pricing_result
    else:
        df_total = pd.concat([df_total, df_pricing_result])

df_total = df_total.reset_index()

df_total.head(3)

df_total.tail(3)

df_final = pd.concat([df, df_total], axis=1)
df_final

df_final.drop('index', inplace=True, axis=1)
df_final.head(3)

resultFilePath = os.path.join(OUT_DIR, 'tiger_divgrowth50_pricing_result.xlsx')
df_final.to_excel(resultFilePath, index=False)

"""
analyze 'tiger_divgrowth50_pricing_result.xlsx'
"""

OUT_DIR = os.path.join(os.getcwd(), 'output')
OUT_DIR

resultFilePath = os.path.join(OUT_DIR, 'tiger_divgrowth50_pricing_result_20230306_by_코드장인.xlsx')
resultFilePath

df_pricing_result = pd.read_excel(resultFilePath)
df_pricing_result.head(3)

df_pricing_result.head(3)

df_pricing_result.sort_values(['매수 점수']).head(10)

df_pricing_result.sort_values(['매수 점수'], ascending=False).head(10)

df_pricing_result.sort_values(['배당률'], ascending=False).head(10)

df_divgrowth_over_7 = df_pricing_result.loc[df_pricing_result['Div Growth Years'] >= 7]
df_divgrowth_over_7

"""
7년 이상 배당 성장주 중 최근 3년 배당 성장률이 7년 성장률보다 높은 종목
"""
df_divgrowth_accelerated = df_divgrowth_over_7.loc[df_divgrowth_over_7['Div Growth Rates(3)']
                                                 > df_divgrowth_over_7['Div Growth Rates(7)']]
df_divgrowth_accelerated

df_divgrowth_accelerated.sort_values(['매수 점수'], ascending=False)

"""
Test
"""

#df_dps = pd.DataFrame(index = range(START_YEAR, END_YEAR+1))
df_dps = pd.DataFrame(index = ('name', range(START_YEAR, END_YEAR+1)))
df_dps.head(3)

df_dps.index.name = 'year'
df_dps
