==> 90_divexcelanalyzer_test.ipynb <==
import kodivstock.divexcelanalyzer as dea

divExcelAnalyzer = dea.DivExcelAnalyzer('./data/tiger_divgrowth50_dps_data_2012_2021.xlsx')

tickers = divExcelAnalyzer.getTickers()
tickers[:3]

divGrowthYears = divExcelAnalyzer.getDivGrowthYears(tickers)
divGrowthYears[:3]

divGrowthRatesFor7Years = divExcelAnalyzer.getDivGrowthRates(tickers, 7)
divGrowthRatesFor7Years[:3]

divGrowthRatesFor3Years = divExcelAnalyzer.getDivGrowthRates(tickers, 3)
divGrowthRatesFor3Years[:3]
