==> 90_pykrx_test.ipynb <==
from pykrx import stock

kospi_list = stock.get_market_ticker_list(market='KOSPI')

kospi_list

len(kospi_list)

kospi_df = stock.get_market_cap('20230203',market='KOSPI')

kospi_df.info()

kospi_df.head(3)

kospi_df.index

ticker = '005930'

from datetime import datetime
from datetime import timedelta

today = datetime.now()
today

end_date = today.strftime("%Y%m%d")
end_date

start_date = (today - timedelta(days=10)).strftime("%Y%m%d")
start_date

end_date = today

df_fundamental = stock.get_market_fundamental(start_date, end_date, ticker)
df_fundamental

df_fundamental.mean()

df_price = stock.get_market_ohlcv(start_date, end_date, ticker)
df_price

round(df_price.mean()['종가'])
