==> 91_web_crawling_test.ipynb <==
https://bigdata-doctrine.tistory.com/3

#리스트에 코스피에 상장된 주식들의 티커 저장하기
# 모듈 가져오기
!pip install pykrx
from pykrx import stock

# 오늘날 코스피에 상장되어 있는 주식의 이름과 티커 수집 
today = datetime.today().strftime("%Y%m%d")
ticker_list = stock.get_market_ticker_list(date = today, market="KOSPI")
symbol_list = []
for ticker in ticker_list:
    symbol = stock.get_market_ticker_name(ticker)
    symbol_list.append(symbol)

# 코스피 상장 주식의 개수 확인
print(len(ticker_list))
print(len(symbol_list))

import requests
from bs4 import BeautifulSoup


ticker_list = ['005930']
symbol_list = ['삼성전자']

# 수집할 증권 데이터의 셀렉터 저장
per_selector = "#_per"
pbr_selector = "#_pbr"
dividend_yield_selector = "#_dvr"


# 크롤링을 사용하여 증권 데이터 수집하기
pers = []
pbrs = []
dividend_yields = []

for ticker in ticker_list:
    url = f"https://finance.naver.com/item/main.nhn?code={ticker}"
    html = requests.get(url, headers={'User-agent':'Mozilla/5.0'})
    soup = BeautifulSoup(html.text, "lxml")
    per = soup.select(per_selector)
    pbr = soup.select(pbr_selector)
    dividend_yield = soup.select(dividend_yield_selector)

    if not per:  # 리스트가 비어있을 경우 None으로 변환
        per = [None]
    for pe in per:
        if pe != None:
            per_text = pe.text
            per_text = per_text.replace(",", "")  # float변환을 위해 따옴표 제거
            pers.append(float(per_text))
        else:
            per_text = pe
            pers.append(per_text)

    if not pbr:  # 리스트가 비어있을 경우 None으로 변환
        pbr = [None]
    for pb in pbr:
        if pb != None:
            pbr_text = pb.text
            pbr_text = pbr_text.replace(",", "")  # float변환을 위해 따옴표 제거
            pbrs.append(float(pbr_text))
        else:
            pbr_text = pb
            pbrs.append(pbr_text)
        
    if not dividend_yield:  # 리스트가 비어있을 경우 None으로 변환
        dividend_yield = [None]
    for d in dividend_yield:
        if d != None:
            dividend_yield_text = d.text
            dividend_yield_text = dividend_yield_text.replace(",", "")  # float변환을 위해 따옴표 제거
            dividend_yields.append(float(dividend_yield_text))
        else:
            dividend_yield_text = d
            dividend_yields.append(dividend_yield_text)

pers

pbrs

dividend_yields
